--- samples/ChatRoomWithAck/.gitignore
@@ -1,4 +1,0 @@
-bin/
-obj/
-.vs/
-**.csproj.user
