--- samples/ChatRoomWithAck/ChatRoomWithAck.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <UserSecretsId>userchatroom</UserSecretsId>
-    <RootNamespace>Microsoft.Azure.SignalR.Samples.ChatRoomWithAck</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="wwwroot\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" Version="2.2.*" />
-    <PackageReference Include="Microsoft.Azure.SignalR" Version="1.0.*" />
-  </ItemGroup>
-</Project>
