--- samples/ChatRoomWithAck/appsettings.json
@@ -1,15 +1,0 @@
-﻿{
-  "Logging": {
-    "IncludeScopes": false,
-    "Debug": {
-      "LogLevel": {
-        "Default": "Debug"
-      }
-    },
-    "Console": {
-      "LogLevel": {
-        "Default": "Debug"
-      }
-    }
-  }
-}
