+++ samples/ReliableChatRoom/.gitignore
@@ -2,4 +2,3 @@ bin/
 obj/
 .vs/
 **.csproj.user
-*.sln
