+++ samples/ReliableChatRoom/Entities/AckInfo.cs
