+++ samples/ReliableChatRoom/Entities/AckResult.cs
