+++ samples/ReliableChatRoom/Entities/Constants.cs
