+++ samples/ReliableChatRoom/Entities/IAckHandler.cs
