+++ samples/ReliableChatRoom/Entities/IMessageHandler.cs
