+++ samples/ReliableChatRoom/Entities/Message.cs
