+++ samples/ReliableChatRoom/Entities/StaticMessageStorage.cs
