+++ samples/ReliableChatRoom/Hub/ReliableRoamingChatroom.cs
