+++ samples/ReliableChatRoom/ReliableChatRoom.csproj
@@ -1,8 +1,8 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
     <TargetFramework>netcoreapp2.2</TargetFramework>
-    <UserSecretsId>reliablechatroom</UserSecretsId>
-    <RootNamespace>Microsoft.Azure.SignalR.Samples.ReliableChatRoom</RootNamespace>
+    <UserSecretsId>userchatroom</UserSecretsId>
+    <RootNamespace>Microsoft.Azure.SignalR.Samples.ChatRoomWithAck</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
